backend/requirements.txt
alembic==1.4.3
astroid==2.4.2
autopep8==1.5.4
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
colorama==0.4.3
contextvars==2.4
dataclasses==0.6
dnspython==2.0.0
email-validator==1.1.1
fastapi==0.61.1
FastAPI-SQLAlchemy==0.1.0
fastapi-utils==0.2.1
geographiclib==1.50
geopy==2.0.0
googlemaps==4.4.2
h11==0.11.0
idna==2.10
immutables==0.14
importlib-metadata==2.0.0
isort==5.6.3
lazy-object-proxy==1.4.3
Mako==1.1.3
MarkupSafe==1.1.1
mccabe==0.6.1
mypy==0.790
mypy-extensions==0.4.3
passlib==1.7.4
pipdeptree==1.0.0
psycopg2==2.8.6
pycodestyle==2.6.0
pydantic==1.6.1
pyflakes==2.2.0
PyJWT==1.7.1
pylint==2.6.0
python-dateutil==2.8.1
python-dotenv==0.14.0
python-editor==1.0.4
requests==2.24.0
six==1.15.0
SQLAlchemy==1.3.19
starlette==0.13.6
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.25.11
uvicorn==0.12.1
wrapt==1.12.1
zipp==3.3.0
